# docs/swagger.yaml
basePath: /
definitions:
  genprotos.Education:
    properties:
      degree:
        type: string
      end_date:
        type: string
      field_of_study:
        type: string
      id:
        type: string
      institution:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  genprotos.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  genprotos.GetAllEducations:
    properties:
      educations:
        items:
          $ref: '#/definitions/genprotos.Education'
        type: array
    type: object
  genprotos.GetAllExperiences:
    properties:
      experiences:
        items:
          $ref: '#/definitions/genprotos.Experience'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Portfolio
  version: "1.0"
paths:
  /education/byuser/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve educatioin by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get Education by User ID Successful
          schema:
            items:
              $ref: '#/definitions/genprotos.Education'
            type: array
        "400":
          description: User ID is required
          schema:
            type: string
        "404":
          description: Educations not found
          schema:
            type: string
        "500":
          description: Error while retrieving educations
          schema:
            type: string
      summary: Get Education by User ID
      tags:
      - Education
  /education/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Education'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Porfolio
      tags:
      - Education
  /education/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Education ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Education
      tags:
      - Education
  /education/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      parameters:
      - in: query
        name: degree
        type: string
      - in: query
        name: end_date
        type: string
      - in: query
        name: field_of_study
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: institution
        type: string
      - in: query
        name: start_date
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllEducations'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Education
      tags:
      - Education
  /education/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Education ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Education'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Education
      tags:
      - Education
  /education/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Education ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Education'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Education
      tags:
      - Education
  /experience/byuser/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve experience by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get Experience by User ID Successful
          schema:
            items:
              $ref: '#/definitions/genprotos.Experience'
            type: array
        "400":
          description: User ID is required
          schema:
            type: string
        "404":
          description: Experiences not found
          schema:
            type: string
        "500":
          description: Error while retrieving experiences
          schema:
            type: string
      summary: Get Experience by User ID
      tags:
      - Experience
  /experience/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Experience'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Experience
      tags:
      - Experience
  /experience/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Experience ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Experience
      tags:
      - Experience
  /experience/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      parameters:
      - in: query
        name: company
        type: string
      - in: query
        name: description
        type: string
      - in: query
        name: end_date
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: start_date
        type: string
      - in: query
        name: title
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllExperiences'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Experience
      tags:
      - Experience
  /experience/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Experience ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Experience'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Experience
      tags:
      - Experience
  /experience/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Experience ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Experience'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Experience
      tags:
      - Experience
...
swagger: "2.0"